rename to endpoints 1: rename directory
