+++ .pre-commit-config.yaml
@@ -61,7 +61,7 @@ repos:
     hooks:
     -   id: endpoints-md
         name: endpoints-md
-        entry: python endpoints.py
+        entry: python endpoints-md.py
         language: system
         types: [python]
         fail_fast: true
