+++ vmngclient/endpoints/configuration_device_software_update.py
@@ -1,4 +1,4 @@
-from vmngclient.primitives import APIPrimitiveBase, post, request
+from vmngclient.endpoints import APIPrimitiveBase, post, request
 from vmngclient.utils.upgrades_helper import SoftwarePackageUpdatePayload
