+++ vmngclient/endpoints/primitive_container.py
@@ -2,15 +2,15 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from vmngclient.primitives.administration_user_and_group import AdministrationUserAndGroupPrimitives
-from vmngclient.primitives.client import ClientPrimitives
-from vmngclient.primitives.configuration_dashboard_status import ConfigurationDashboardStatusPrimitives
-from vmngclient.primitives.configuration_device_software_update import ConfigurationDeviceSoftwareUpdatePrimitives
-from vmngclient.primitives.configuration_device_template import ConfigurationDeviceTemplatePrimitives
-from vmngclient.primitives.monitoring_device_details import MonitoringDeviceDetailsPrimitives
-from vmngclient.primitives.tenant_backup_restore import TenantBackupRestorePrimitives
-from vmngclient.primitives.tenant_management import TenantManagementPrimitives
-from vmngclient.primitives.tenant_migration import TenantMigrationPrimitives
+from vmngclient.endpoints.administration_user_and_group import AdministrationUserAndGroupPrimitives
+from vmngclient.endpoints.client import ClientPrimitives
+from vmngclient.endpoints.configuration_dashboard_status import ConfigurationDashboardStatusPrimitives
+from vmngclient.endpoints.configuration_device_software_update import ConfigurationDeviceSoftwareUpdatePrimitives
+from vmngclient.endpoints.configuration_device_template import ConfigurationDeviceTemplatePrimitives
+from vmngclient.endpoints.monitoring_device_details import MonitoringDeviceDetailsPrimitives
+from vmngclient.endpoints.tenant_backup_restore import TenantBackupRestorePrimitives
+from vmngclient.endpoints.tenant_management import TenantManagementPrimitives
+from vmngclient.endpoints.tenant_migration import TenantMigrationPrimitives
 
 if TYPE_CHECKING:
     from vmngclient.session import vManageSession
